requirements.txt
requests
python-dotenv
serpapi
python-docx
langchain
langgraph
langchain-core
langchain-community
langchain-google-genai
langchain-chroma
ipykernel
rich
markitdown[all]
pytest
pytest-asyncio
pytest-mock
pytest-icdiff
streamlit
altair
Authlib
streamlit-tags
pytest-asyncio
respx
rapidfuzz
streamlit-local-storage
google-genai
langchain-openai
langchain-anthropic
langchain-pinecone
